requirements.txt
requests>=2.32.0
pillow>=10.0.0
langchain>=1.0.0
langchain-openai>=1.0.0
langgraph>=1.0.0
